split modifier types
